feat: Add BitLife-like systems and fix critical issues
- Fix Age property to allow setting (was read-only)
- Convert decimal types to float for Unity JsonUtility compatibility
- Fix HashSet FIFO issue with Queue-based implementation

New Systems:
- ActivitiesSystem: Gym, meditation, doctor, library, movies, social media, vacation
- CareerSystem: Job listings, applications, promotions, raises, yearly salary
- RelationshipSystem: Talk, gift, spend time, compliment, argue, reconcile, flirt, propose, divorce
- CrimeSystem: Shoplift, car theft, bank robbery, scam, drugs, assault

Improvements:
- Add Fame stat to UI (was missing)
- Add NPC aging system (NPCs now age and can die)
- Add 50+ new events (school, health, career, social, romance, family, financial)

+++ Assets/Scripts/Core/EventBus.cs
@@ -164,9 +164,9 @@ namespace TurkishLifeSim.Core
     /// </summary>
     public struct MoneyChangedEvent : IGameEvent
     {
-        public decimal OldAmount;
-        public decimal NewAmount;
-        public decimal Delta;
+        public float OldAmount;
+        public float NewAmount;
+        public float Delta;
         public string Reason;
     }
